Moving logic for catching the visible pokemon out into a new worker
