+++ pokemongo_bot/cell_workers/__init__.py
@@ -5,3 +5,4 @@ from seen_fort_worker import SeenFortWorker
 from move_to_fort_worker import MoveToFortWorker
 from initial_transfer_worker import InitialTransferWorker
 from evolve_all_worker import EvolveAllWorker
+from catch_visible_pokmeon_worker import CatchVisiblePokemonWorker
